Update read.me
just adding a few lines to the process that I had to do to be able to run config:publish or migrate:publish. The reason being that the source files hadn't been brought in yet for your package at this point, and it would be impossible to publish without your source code there in the vendor. 

Hope it helps!